ocr-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
paddleocr==2.7.3
paddlepaddle==2.6.0
pdf2image==1.17.0
Pillow==10.2.0
